requirements.txt
openai>=1.35.0
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.66.0